Reorganize routes: run stuff in workflow to run routes, docker stuff to a utils file
